Explicit boxing
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/lang/trunk@1391027 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/lang3/time/FormatCache.java
@@ -141,13 +141,13 @@ abstract class FormatCache<F extends Format> {
             try {
                 DateFormat formatter;
                 if (dateStyle == null) {
-                    formatter = DateFormat.getTimeInstance(timeStyle, locale);                    
+                    formatter = DateFormat.getTimeInstance(timeStyle.intValue(), locale);                    
                 }
                 else if (timeStyle == null) {
-                    formatter = DateFormat.getDateInstance(dateStyle, locale);                    
+                    formatter = DateFormat.getDateInstance(dateStyle.intValue(), locale);                    
                 }
                 else {
-                    formatter = DateFormat.getDateTimeInstance(dateStyle, timeStyle, locale);
+                    formatter = DateFormat.getDateTimeInstance(dateStyle.intValue(), timeStyle.intValue(), locale);
                 }
                 pattern = ((SimpleDateFormat)formatter).toPattern();
                 String previous = cDateTimeInstanceCache.putIfAbsent(key, pattern);
